k8swatcher: simplify resource version handling logic
Parsing the resourceVersion as an integer is strictly prohibited by the
Kubernetes API [1]. During a review of code that would be affected by
non-numerical resourceVersions, I came upon this code. It looks like
parsing of resourceVersions was only done to detect out-of-order events
from the watch stream, but Kubernetes watches are guaranteed to deliver
events in order, so this parsing logic is not necessary.

[1] https://kubernetes.io/docs/reference/using-api/api-concepts/#resou

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>